+++ Examples/ParkingSystem/PaymentStrategies/TransactionStrategy.java
@@ -2,5 +2,5 @@ package Examples.ParkingSystem.PaymentStrategies;
 
 public interface TransactionStrategy {
     boolean validatePaymentDetails();
-    boolean transact(Integer amount);
+    boolean transact(Double amount);
 }
